+++ src/schemas/v1/api/metrics.rs
@@ -1,5 +1,5 @@
 use crate::schemas::v1::db::queries::{self as db};
-use crate::schemas::v1::db::tables;
+use crate::models::metrics::Metric;
 use rocket::{delete, get, post, serde::json::Json, State};
 use sqlx::MySql;
 
@@ -7,7 +7,7 @@ use sqlx::MySql;
 pub async fn get_metrics_by_app_id(
     pool: &State<sqlx::Pool<MySql>>,
     instance_id: Option<i64>,
-) -> Json<Vec<tables::Metric>> {
+) -> Json<Vec<Metric>> {
     let instance_id = instance_id.or(Some(0)); // Set to 0 (or null equivalent) if blank
     let metrics = db::metrics::get_metrics_by_app_id(pool, instance_id)
         .await
@@ -17,7 +17,7 @@ pub async fn get_metrics_by_app_id(
 }
 
 #[get("/metrics")]
-pub async fn get_metrics(pool: &State<sqlx::Pool<MySql>>) -> Json<Vec<tables::Metric>> {
+pub async fn get_metrics(pool: &State<sqlx::Pool<MySql>>) -> Json<Vec<Metric>> {
     let metrics = db::metrics::get_metrics_by_app_id(pool, None)
         .await
         .unwrap();
