+++ src/schemas/v1/api/providers.rs
@@ -5,11 +5,11 @@
 //! It is designed to be used primarily by the dashboard to add new providers and manage existing ones.
 //! The resulting database table is read at runtime by the various directories to determine which provider configs they need to have access to.
 
-use crate::db::tables::ProviderAuditLog;
-use crate::schemas::v1::db::tables::{Provider, Instance};
+use crate::models::provider::{ProviderAuditLog, Provider};
 use crate::schemas::v1::db::queries::{self as queries};
-use rocket::http::Status;
 use rocket::serde::json::{Json, Value};
+use crate::models::instance::Instance;
+use rocket::http::Status;
 use serde_json::json;
 
 /// List all providers in the system with pagination support.
