+++ src/schemas/v1/db/queries/metrics.rs
@@ -1,5 +1,5 @@
 use sqlx::{FromRow, MySql, Pool};
-use crate::db::tables::Metric;
+use crate::models::metrics::Metric;
 
 pub async fn get_metrics_by_app_id(pool: &Pool<MySql>, app_id: Option<i64>) -> anyhow::Result<Vec<Metric>> {
     let query = if let Some(app_id) = app_id {
