+++ src/schemas/v1/models/mod.rs
@@ -16,3 +16,5 @@ pub mod provider;
 pub mod storage;
 pub mod notification;
 pub mod cost;
+pub mod util_tables;
+pub mod role;
